Add some e2e tests (#26)
* Add some initial tests

* fix publish

* fix publish 2

* fix issue

* fix fleet tests

* add pod perms

* fix image pull secret

* add secret unit tests

* fix tests....

* fix fleet

* fix fleet

* Fix: ProdChecker used

* Add test, fix issues, refactor

* Remove timeout

* Fix test?

* Fix test??

* Fix test again please

* Refactor tests slightly

---------

+++ operator/internal/controller/fleet_controller.go
@@ -240,6 +240,9 @@ func (r *FleetReconciler) handleDeletion(ctx context.Context, fleet *gameserverv
 		return err
 	}
 	for _, server := range servers.Items {
+		if server.DeletionTimestamp != nil {
+			continue
+		}
 		if err := r.Delete(ctx, &server); err != nil {
 			return err
 		}
